Unsupervised learning to select candidates for a job/internship profile
This python code uses clustering and PCA to group candidates to find the top candidates with the highest information (skills) provided in their application. But the most challenging and important part was to prepare the data to make sense of it. 
The selected candidates are shown towards the end of the codes.
This project was completed in little over 24 hours.